requirements.txt
requests
pyyaml
python-dotenv
networkx
matplotlib
numpy
